backend/pyproject.toml
[project]
name = "explanatory-note"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "docxtpl>=0.20.2",
    "fastapi>=0.128.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "num2words>=0.5.14",
    "PyJWT>=2.9.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.13.1",
    "httpx>=0.28.1",
    "alembic>=1.18.4",
    "psycopg[binary]>=3.3.3",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
